Access to backdrop :on-click event
I become *very angry* when I website won't close the modal when I click on the backdrop.
I wouldn't' dare do that to any of my users.

By adding the :backdrop-on-click, using a function to close the modal is trivial.

Ultimately, it should also be possible to close the modal with `esc`. What would be the best way to do that? Adding and removing a listener on `component-did-mount` `component-will-unmount`?