Retain-Focus-On-Enter-In-Typeahead
After selecting an option from the list of values after typing in a desired string value, the focus is lost and users probably have to use the mouse to point to the typeahead input field again or use the combination of TAB and SHIFT+TAB to regain focus. This is a small feature I wanted to propose to save some time.