Several fixes for typeahead
Related to: https://github.com/Day8/re-com/issues/190

Trying to use the typeahead was causing us a world of
headaches. Below is a list of specific issues that we were facing that
this PR addresses. With these new changes, typeahead appears to work
fine with any combination of `change-on-blur?`, `rigid?`, & `immediate-model-update?`.

- Tabbing off an element with no changes broke the entire typeahead
  so it would never work again
- Tabbing off an element with a change just produced an error because
  args were in the wrong order in the fired function call
- Once the tabbing issues were fixed, we couldn't use the typeahead
  element after an initial selection (it was firing the update event
  on every keypress while we were trying to get the typeahead list).

Whenever a user tries to tab out of a typeahead box, if the 2
conditions in the `and` of `re-com.typeahead/input-text-will-blur`
aren't met, then the local state-atom for typeahead is replaced with
`nil` instead of having its value updated. This causes the entire
component to just break.

Blurring off when there was an update was calling `update-model` with
the args in the wrong order.

Nothing in the code was actually updating the external value
(i.e. `external-model`) stored in the local state, which means on any
text input after an update was firing the model update, which
prevented new suggestions from being rendered (since the local state
was constantly being updated to the previously stored value). This
update/check seems completely unnecessary, since the code already
updates the model on suggestion changes anyways.

Blurring off the text box doesn't need to trigger `got-suggestions`,
we've already done that with other key events, instead it should
behave the same as tabbing through the event (i.e. they're both ways
to get to the blur, they should behave the same).