Change component argument validation to not throw on error
This significantly cleans up the console when an argument validation
error occurs. Instead of throwing, we just use js/console.error to print
a more succinct error. This also means that apps with argument
validation errors will not get a 'white screen of death' on the uncaught
exception during development which is probably a good thing for e.g. hot
code reloading productivity. The only downside is errors may be harder
to notice.