First batch of changes requested in PR
rename [only-button?] to [show-only-button?] in code and in parameters for selection list and tag-dropdown
updating tag-dropdown to have show-only-button? instead of only-button? update docs too
change all paramter passing internally to keyword arguments (within reason)
rename group-componet to filter-group
add test for test-remove-item-with-cleanup
change the value-entry-box to a defmulti/defmethod setup from a big cond and reduce code repetition
adjust the re-com/dropdown.cljs dropdown component to more properly display it is in the disabled? state
remove show-border and top-label
change from the janky JS buttons to the re-com dropdowns