Validate required parts
Now, we can declare :part-required? in the part-structure. That way,
a hiccup will pass validation as long as that key is declared within
`:parts`.

If the part also has `:top-level-arg?` declared, then it can also pass
validation when declared within the top-level args of a hiccup.

Prior to the parts/theme update, some parameters were declared with
`:required true`. For instance, the `:label` parameter to
`re-com.buttons/button`. Since `:label` is now a part, that validation
should no longer apply. That means we have to remove `:required true`
from any parameters that we've turned into parts.